requirements.txt
Flask
Flask-SQLAlchemy
																																																																																						